backend/requirements.txt
fastapi
uvicorn
requests
pydantic
supabase
httpx
python-dotenv
pandas
